create at_internal.h add AT_Scene and AT_Model to it
